requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
mysql-connector-python==8.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
asgiref==3.7.2
gunicorn==21.2.0